+++ snowflake-distributed-service-impl/src/main/java/com/dwarfeng/sfds/impl/handler/GenerateHandlerImpl.java
@@ -78,11 +78,11 @@ public class GenerateHandlerImpl implements GenerateHandler {
      * 开始时间截（毫秒）。
      * <p>
      * 用于计算生成 ID 中的时间戳部分，计算公式为：生成 ID 中的时间戳部分 = (当前时间戳 - twepoch) << 22
-     * 如果未配置此项，将使用 {@link SnowflakeConstants#TWEPOCH} 的默认值。
+     * 如果未配置此项，将使用 {@link SnowflakeConstants#DEFAULT_TWEPOCH} 的默认值。
      */
     // 该单词系专有术语，忽略相关的拼写警告。
     @SuppressWarnings({"SpellCheckingInspection", "RedundantSuppression"})
-    @Value("${snowflake.twepoch:#{T(com.dwarfeng.sfds.sdk.util.SnowflakeConstants).TWEPOCH}}")
+    @Value("${snowflake.twepoch:#{T(com.dwarfeng.sfds.sdk.util.SnowflakeConstants).DEFAULT_TWEPOCH}}")
     private long twepoch;
 
     /**